Change first parameter of Vector4i.mul(float, Vector4i) to int (#166)

+++ src/org/joml/Vector4i.java
@@ -669,18 +669,18 @@ public class Vector4i implements Externalizable, Vector4ic {
      *          the scalar to multiply by
      * @return a vector holding the result
      */
-    public Vector4i mul(float scalar) {
+    public Vector4i mul(int scalar) {
         return mul(scalar, thisOrNew());
     }
 
     /* (non-Javadoc)
-     * @see org.joml.Vector4ic#mul(float, org.joml.Vector4i)
+     * @see org.joml.Vector4ic#mul(int, org.joml.Vector4i)
      */
-    public Vector4i mul(float scalar, Vector4i dest) {
-        dest.x = (int) (x * scalar);
-        dest.y = (int) (y * scalar);
-        dest.z = (int) (z * scalar);
-        dest.w = (int) (w * scalar);
+    public Vector4i mul(int scalar, Vector4i dest) {
+        dest.x = x * scalar;
+        dest.y = y * scalar;
+        dest.z = z * scalar;
+        dest.w = w * scalar;
         return dest;
     }
